Tried to separate the 5 tabs: overview and the 4 kanbans, into their own componenet files
