+++ app/dashboard/components/shared.tsx
@@ -1,0 +1,51 @@
+'use client';
+
+import { useDroppable } from '@dnd-kit/core';
+
+// Shared components used across multiple tabs
+
+export function CardDateMeta({
+  created,
+  completed,
+  className,
+}: {
+  created?: string | null;
+  completed?: string | null;
+  className?: string;
+}) {
+  const formatCardDate = (value?: string | null) => {
+    if (!value) return '-';
+    try {
+      const date = new Date(value);
+      if (Number.isNaN(date.getTime())) return '-';
+      return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' });
+    } catch {
+      return '-';
+    }
+  };
+
+  return (
+    <div className={className ? className : 'mt-3'}>
+      <div className="flex items-center justify-between text-[10px] uppercase tracking-wider text-gray-400 mb-1">
+        <span>Created:</span>
+        <span>Completed:</span>
+      </div>
+      <div className="flex items-center justify-between text-xs text-yellow-400">
+        <span>{formatCardDate(created)}</span>
+        <span>{formatCardDate(completed)}</span>
+      </div>
+    </div>
+  );
+}
+
+export function DroppableColumn(props: { id: string; children: React.ReactNode }) {
+  const { setNodeRef, isOver } = useDroppable({ id: props.id });
+  return (
+    <div ref={setNodeRef} className={`space-y-3 min-h-32 ${isOver ? 'outline outline-2 outline-electric-blue/60 outline-offset-2 bg-gray-650/40' : ''}`}>
+      {props.children}
+      {/* When empty, provide space to drop */}
+      <div className="h-2"></div>
+    </div>
+  );
+}
+
